// src/ui-types.d.ts
export interface TableContent {
  jobTitle: string;
  companyName: string;
  location: string;
  appStatus: string;
  jobType?: string;
  appDate?: string;
  method?: string;
  description?: string;
  url?: string;
  jobSalary?: string;
  companyContact?: string;
  interviewDate?: string;
  interviewNotes?: string;
  followUp?: string;
  addtionalNotes?: string;
}